Flatten test dir.
I don't know why. I'm tired of clicking through directory trees, I guess.
