+++ t/811_fault_injection.c
